do not unlink tempfiles if it was not created

+++ boto_addins/s3.py
@@ -247,7 +247,8 @@ def async_http_download(source, destination,
             )
 
     except BaseException:
-        os.unlink(tmp_name)
+        if os.path.exists(tmp_name):
+            os.unlink(tmp_name)
         raise
 
     # concurrent downloading
